fix link component, create file interface, remove file App.css, move file App.tsx
