+++ src/app/pages/Home/partials/AboutItem.tsx
@@ -1,22 +1,9 @@
 import React from 'react';
+import { IAbout } from '../../../shared/interfaces/about';
 import Badge from './../../../shared/components/partials/Badge';
-import Button from './../../../shared/components/partials/Button';
+import Link from './../../../shared/components/partials/Link';
 
-interface IAboutItemProps {
-  title: string;
-  classCol: string;
-  classIndex: string;
-  link: {
-    title: string;
-    customClass: string
-  };
-  badge?: {
-    discount: number;
-    type: string;
-  };
-}
-
-const AboutItem: React.FC<IAboutItemProps> = ({
+const AboutItem: React.FC<IAbout> = ({
   title,
   classCol,
   classIndex,
@@ -34,9 +21,10 @@ const AboutItem: React.FC<IAboutItemProps> = ({
           )}
           <h4 className="typo-1 about-title">{title}</h4>
           <div className="about-link">
-            <Button
+            <Link
               title={link.title}
               customClass={link.customClass}
+              path={link.path}
             />
           </div>
         </div>
