+++ src/app/shared/components/partials/Link.tsx
@@ -3,11 +3,12 @@ import React from 'react';
 interface ILinkProps {
   title: string,
   customClass: string,
+  path: string
 }
 
-const Link: React.FC<ILinkProps> = ({ title, customClass }) => {
+const Link: React.FC<ILinkProps> = ({ title, customClass, path }) => {
   return (
-    <a href="#" className={customClass ? `btn ${customClass}`: 'btn'}>
+    <a href={path} className={customClass}>
       {title}
     </a>
   )
