+++ src/app/shared/interfaces/chooseus.ts
@@ -1,0 +1,5 @@
+export interface IChooseus {
+  name: string;
+  icon: string;
+  desc: string;
+}
