+++ src/app/shared/interfaces/product.ts
